requirements.txt
torch
numpy
h5py
tqdm
wandb
hydra-core
omegaconf